Despiece

+++ mod_repa/tablas/productos/productos_delete.php
@@ -20,7 +20,7 @@
 
         if($perfilSirep == 1){
 
-            $query0         = "SELECT foto FROM rep3_productos WHERE producto_id = '{$id}'";
+            $query0         = "SELECT foto, foto_despiece FROM rep3_productos WHERE producto_id = '{$id}'";
             $sentenciaSQL   = $conexion->prepare($query0);
             $sentenciaSQL   -> execute();
             $respuesta0     = $sentenciaSQL->fetch(PDO::FETCH_ASSOC);
@@ -31,8 +31,14 @@
             
             if($sentenciaSQL->rowCount() > 0){
                 $arrayRespuesta['estado'] = "Transacción exitosa";
-                chdir('./adjuntos');
-                unlink($respuesta0['foto']);
+                if($respuesta0['foto'] != ''){
+                    chdir('./adjuntos');
+                    unlink($respuesta0['foto']);
+                }
+                if($respuesta0['foto_despiece'] != ''){
+                    chdir('./adjuntos');
+                    unlink($respuesta0['foto_despiece']);
+                }
             } else {
                 $arrayRespuesta['estado'] = "Algo salió mal";
             }
